mdadm: Fix mdadm -r remove option regresision
The commit noted below globally adds a parameter to the -r option but missed
the fact that -r is used for another purpose: --remove.

After that commit a command such as:

  mdadm /dev/md0 -r /dev/loop0

will do nothing seeing the device parameter will be consumed as a
argument to the -r option; thus there will only be one device
seen one the command line, devs_found will only be 1 and nothing will
happen.

This caused the 01r5integ and 01raid6integ tests to hang indefinitely
as mdadm did not remove the failed device. With the device not removed,
it would not be readded. Then the loop waiting for the array status to
change would loop forever.

To fix this, revert the changes in the noted patch and create a new subopt
type for the monitor mode with parameters required for -r.

Fixes: 546047688e1c ("mdadm: fix coredump of mdadm --monitor -r")
Signed-off-by: Logan Gunthorpe <logang@deltatee.com>
Cc: Wu Guanghao <wuguanghao3@huawei.com>
Cc: Mariusz Tkaczyk <mariusz.tkaczyk@linux.intel.com>